Improve code modularity
Extract SQLite logic and BibTeX escape sequence substitution logic
to dedicated scripts that can later be unit tested.
